Rename model file as class
