+++ src/class.ts
